CXP-1100: Move ExternalApi to Public folder
