+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Apps/Public/AuthorizeEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Apps\ExternalApi;
+namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Apps\Public;
 
 use Akeneo\Connectivity\Connection\Application\Apps\Command\CreateAppWithAuthorizationCommand;
 use Akeneo\Connectivity\Connection\Application\Apps\Command\CreateAppWithAuthorizationHandler;
