+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Apps/Public/RequestAccessTokenActionEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Apps\ExternalApi;
+namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Apps\Public;
 
 use Akeneo\Connectivity\Connection\Application\Apps\Command\GenerateAsymmetricKeysCommand;
 use Akeneo\Connectivity\Connection\Application\Apps\Command\GenerateAsymmetricKeysHandler;
